exclude the SBT Scala compiler ClassLoader from instrumentation
