+++ agent/src/main/java/kanela/agent/util/classloader/ClassLoaderNameMatcher.java
@@ -56,6 +56,10 @@ public class ClassLoaderNameMatcher extends ElementMatcher.Junction.AbstractBase
         return new ClassLoaderNameMatcher("sbt.internal.PluginManagement$PluginClassLoader");
     }
 
+    public static ElementMatcher.Junction.AbstractBase<ClassLoader> isSBTCompilerClassLoader() {
+        return new ScalaCompilerClassLoaderMatcher();
+    }
+
     public static ElementMatcher.Junction.AbstractBase<ClassLoader> isKanelaClassLoader() {
         return new ClassLoaderNameMatcher(ChildFirstURLClassLoader.class.getName());
     }
